Change text_color to background_color
Originally I wasn't thinking and just assumed the magic number was text color, don't ask me why. As it turns out,
'#e0e0e0' is a light gray and '#111111' is nearly black. Using this and the context of the is_dark_mode flag, I think
it makes the most sense to call this background color. This is also a good example about why magic numbers are bad